# Optimize quad rendering by grouping quads