--- a/data/shader/quad.frag
+++ b/data/shader/quad.frag
@@ -2,10 +2,18 @@
 uniform sampler2D gTextureSampler;
 #endif
 
+#ifndef TW_QUAD_GROUPED
 uniform vec4 gVertColors[TW_MAX_QUADS];
+#else
+uniform vec4 gVertColors[1];
+#endif
 
 noperspective in vec4 QuadColor;
+#ifndef TW_QUAD_GROUPED
 flat in int QuadIndex;
+#else
+#define QuadIndex 0
+#endif
 #ifdef TW_QUAD_TEXTURED
 noperspective in vec2 TexCoord;
 #endif