… src/engine/client/backend/vulkan/backend_vulkan.cpp