<!-- What is the motivation for the changes of this pull request? -->

<!-- Note that builds and other checks will be run for your change. Don't feel intimidated by failures in some of the checks. If you can't resolve them yourself, experienced devs can also resolve them before merging your pull request. -->

With the pipeline in place we now after setup, if any quads in a quadlayer are animated or not. If none of them are animated, we can optimize their rendering by pushing their location and color.

This only applies to vulkan and OpenGL 3.3, falls back to default rendering in OpenGL 3 and 1

## Vulkan:
Before ~160 FPS:
![screenshot_2025-06-10_15-14-48](https://github.com/user-attachments/assets/8872a20f-7ee8-4525-9180-cb52b55413d5)

After >1300 FPS (depends on zoom level):
![screenshot_2025-06-11_12-02-48](https://github.com/user-attachments/assets/288d992c-008e-4619-bfbc-4e1e4d680c4a)

## OpenGL 3.3
Before ~36 FPS (no screenshot)

After ~ 1100 FPS:
![screenshot_2025-06-11_16-08-02](https://github.com/user-attachments/assets/452dbcc1-5b9f-45ad-8e2c-a00a55d78c54)


@Jupeyy 

## Checklist

- [x] Tested the change ingame
- [x] Provided screenshots if it is a visual change
- [x] Tested in combination with possibly related configuration options
- [ ] Written a unit test (especially base/) or added coverage to integration test
- [ ] Considered possible null pointers and out of bounds array indexing
- [ ] Changed no physics that affect existing maps
- [ ] Tested the change with [ASan+UBSan or valgrind's memcheck](https://github.com/ddnet/ddnet/#using-addresssanitizer--undefinedbehavioursanitizer-or-valgrinds-memcheck) (optional)