@AssassinTee 2025-06-13
> The new shader programs are not being deleted yet.

done, program gets deleted and pointer freed